pyproject.toml
[project]
name = "dakora-examples"
version = "0.1.0"
description = "Standalone examples for Dakora instrumentation and templates."
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["python-dotenv>=1.0.0"]

[project.optional-dependencies]
core = ["dakora>=0.1.0", "dakora-instrumentation>=0.2.0", "python-dotenv>=1.0.0"]
otel = ["opentelemetry-sdk>=1.24.0", "opentelemetry-exporter-otlp-proto-http>=1.24.0"]
openai = ["openai>=1.0.0", "opentelemetry-instrumentation-openai>=0.1.0"]
anthropic = ["anthropic>=0.18.0", "opentelemetry-instrumentation-anthropic>=0.1.0"]
maf = ["agent-framework-core>=1.0.0b251016", "opentelemetry-exporter-otlp-proto-http>=1.24.0"]
jaeger = ["opentelemetry-exporter-jaeger-thrift>=1.24.0"]
quickstart = [
  "dakora>=0.1.0",
  "dakora-instrumentation>=0.2.0",
  "python-dotenv>=1.0.0",
  "opentelemetry-sdk>=1.24.0",
  "opentelemetry-exporter-otlp-proto-http>=1.24.0",
  "opentelemetry-instrumentation-openai>=0.1.0",
  "openai>=1.0.0",
]
all = [
  "dakora>=0.1.0",
  "dakora-instrumentation>=0.2.0",
  "python-dotenv>=1.0.0",
  "opentelemetry-sdk>=1.24.0",
  "opentelemetry-exporter-otlp-proto-http>=1.24.0",
  "opentelemetry-instrumentation-openai>=0.1.0",
  "openai>=1.0.0",
  "anthropic>=0.18.0",
  "opentelemetry-instrumentation-anthropic>=0.1.0",
  "agent-framework-core>=1.0.0b251016",
  "opentelemetry-exporter-jaeger-thrift>=1.24.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
